sepalating file and adding method
